Allow fetch() to be cached in memory
When multiple modules use the same http source, requesting a fresh copy
for each module is an unnecessary load. This change allows fetch() to
accept a third parameter to enable caching. In this case the full HTTP
result will be held in RAM. If the same request with enabled caching is
made again, it will return the cached data instead of contacting the
source.